fix: deleteCast 로직 보완 캐스트 삭제할 때, 이미지url은 기본이미지가 아닐 때만 삭제하도록 변경

+++ src/main/java/com/umc/owncast/domain/cast/service/CastServiceImpl.java
@@ -218,7 +218,9 @@ public class CastServiceImpl implements CastService {
         if(!Objects.equals(cast.getMember(), member)){
             throw new UserHandler(ErrorCode.NO_AUTHORITY);
         }
-        fileService.deleteFile(cast.getImagePath());
+        if(!cast.getImagePath().equals(CAST_DEFAULT_IMAGE_PATH)) {
+            fileService.deleteFile(cast.getImagePath());
+        }
         fileService.deleteFile(cast.getFilePath());
 
         castPlaylistRepository.deleteAllByCast(cast);
